--- src/main/java/voideventhub/voidcore/common/events/EntityLoadEvent.java
@@ -1,18 +1,0 @@
-package voideventhub.voidcore.common.events;
-
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
-import net.minecraft.server.network.ServerPlayerEntity;
-import voideventhub.voidcore.common.PlayerActionService;
-
-public class EntityLoadEvent {
-
-    public static void register() {
-        ServerEntityEvents.ENTITY_LOAD.register((entity, world) -> {
-            if (entity instanceof ServerPlayerEntity player) {
-                PlayerActionService service = new PlayerActionService(world.getServer().getPlayerManager());
-                service.playerJoin(player.getUuid());
-            }
-        });
-    }
-
-}
