+++ src/main/java/voideventhub/voidcore/common/events/VPSEvents.java
@@ -7,16 +7,13 @@ public class VPSEvents {
 
     public static void register() {
         ServerTickEvent.register();
+        PlayerDisconnectEvent.register();
+        PlayerJoinEvent.register();
+        AfterDeathEvent.register();
 
-        // These events make queries to the database, so we only want to register them
-        // if we have a valid connection to avoid crashes
         if (((MongoDbRepository)MongoDbRepository.getInstance()).validConnection()) {
             VoidCore.LOGGER.error("Could not connect to database.");
-            return;
         }
-
-        EntityUnloadEvent.register();
-        EntityLoadEvent.register();
     }
 
 }
